backend/requirements.txt
asgiref==3.8.1
certifi==2025.6.15
charset-normalizer==3.4.2
Django==5.2.3
django-cors-headers==4.7.0
django-debug-toolbar==5.2.0
django-extensions==4.1
djangorestframework==3.16.0
drf-yasg==1.21.10
ecdsa==0.19.1
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
packaging==25.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pydotplus==2.0.2
pyparsing==3.2.3
python-decouple==3.8
python-jose==3.5.0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
rsa==4.9.1
six==1.17.0
sqlparse==0.5.3
uritemplate==4.2.0
urllib3==2.5.0
whitenoise==6.9.0
gunicorn